wip test

+++ src/utils/functions/deleteTempImages.ts
@@ -1,7 +1,6 @@
 import fs from "fs";
 import path from "path";
 
-// const uri = "../public/temporal";
 export const deleteTemporalImages = (dirPath: string) => {
   fs.readdir(dirPath, (err, files) => {
     if (err) {
@@ -10,14 +9,17 @@ export const deleteTemporalImages = (dirPath: string) => {
     }
 
     files.forEach((file) => {
-      const nextPath = path.join(dirPath, file);
-      fs.unlink(nextPath, (err) => {
-        if (err) {
-          console.error("Error deleting file: ", err);
-        } else {
-          console.log(`Deleted file ${nextPath}`);
-        }
-      });
+      if (!file.endsWith(".md")) {
+        const nextPath = path.join(dirPath, file);
+
+        fs.unlink(nextPath, (err) => {
+          if (err) {
+            console.error("Error deleting file: ", err);
+          } else {
+            console.log(`Deleted file ${nextPath}`);
+          }
+        });
+      }
     });
   });
 };
